requirements.txt
streamlit==1.20.0
diffusers==0.14.0
transformers==4.26.0
accelerate==0.18.0
torch==2.0.1
pillow==9.5.0
huggingface_hub==0.14.1
numpy==1.23.5
